requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
sqlalchemy==2.0.23

# Network scanning
python-nmap==0.7.1
scapy==2.5.0
netifaces==0.11.0

# Vulnerability detection
requests==2.31.0
beautifulsoup4==4.12.2

# Scheduling and async
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Alerts and notifications
python-dotenv==1.0.0
emails==0.6
slack-sdk==3.26.1

# Security and validation
cryptography==41.0.7
pydantic==2.5.2
python-jose==3.3.0

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Development (optional, remove for production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0